fix: /today のサーバー側 Phase 1（認証集約・並列化・プリセットキャッシュ）
- react/cache で getSupabaseAuthForRequest を提供し、RSC/Server Actions 内の getUser を同一リクエスト 1 回に集約
- getOrCreateSnapshotRestaurant を Promise.all に含めて初回取得を並列化
- loadPresets を lib/presets-server でプロセス内キャッシュ

Issue #145 Phase 1

Made-with: Cursor
